+++ opr/primer.py
@@ -13,7 +13,7 @@ from .params import PRIMER_ADDITION_ERROR, PRIMER_MULTIPLICATION_ERROR
 from .params import PRIMER_MELTING_TEMPERATURE_NOT_IMPLEMENTED_ERROR
 from .params import PRIMER_ATTRIBUTE_NOT_COMPUTABLE_ERROR
 from .functions import molecular_weight_calc, basic_melting_temperature_calc, salt_adjusted_melting_temperature_calc, gc_clamp_calc
-from .functions import nearest_neighbor_melting_temperature_calc
+from .functions import nearest_neighbor_melting_temperature_calc, calculate_thermodynamics_constants
 from .functions import e260_ssnn_calc, protein_seq_calc
 
 
@@ -59,6 +59,8 @@ class Primer:
             MeltingTemperature.SALT_ADJUSTED: None,
             MeltingTemperature.NEAREST_NEIGHBOR: None,
         }
+        self._delta_s = None
+        self._delta_h = None
 
         # Track computed attributes
         self._computed = {
@@ -73,6 +75,8 @@ class Primer:
                 MeltingTemperature.SALT_ADJUSTED: False,
                 MeltingTemperature.NEAREST_NEIGHBOR: False,
             },
+            "delta_s": False,
+            "delta_h": False,
         }
 
     def is_computed(self, attr):
@@ -333,6 +337,32 @@ class Primer:
             self._computed["E260"] = True
         return self._E260
 
+    @property
+    def delta_s(self):
+        """
+        Calculate ΔS based on Neareset neighbor method.
+
+        :return: ΔS (entropy change, kcal/mol·K)
+        """
+        if not self._computed["delta_s"]:
+            self._delta_h , self._delta_s = calculate_thermodynamics_constants(self._sequence)
+            self._computed["delta_s"] = True
+            self._computed["delta_h"] = True
+        return self._delta_s
+
+    @property
+    def delta_h(self):
+        """
+        Calculate ΔH based on Neareset neighbor method.
+
+        :return: ΔH (enthalpy change, kcal/mol)
+        """
+        if not self._computed["delta_h"]:
+            self._delta_h , self._delta_s = calculate_thermodynamics_constants(self._sequence)
+            self._computed["delta_s"] = True
+            self._computed["delta_h"] = True
+        return self._delta_h
+
     def repeats(self, sequence, consecutive=False):
         """
         Count occurrences of a subsequence in a given sequence.
@@ -372,6 +402,9 @@ class Primer:
         else:
             # the method is MeltingTemperature.NEAREST_NEIGHBOR
             self._melting_temperature[MeltingTemperature.NEAREST_NEIGHBOR] = nearest_neighbor_melting_temperature_calc(
-                self._sequence, self._salt_level)
+                self._sequence,
+                self._salt_level,
+                (self.delta_h, self.delta_s)
+                )
         self._computed["melting_temperature"][method] = True
         return self._melting_temperature[method]