+++ opr/functions.py
@@ -3,8 +3,8 @@
 import math
 from .params import A_WEIGHT, T_WEIGHT, C_WEIGHT, G_WEIGHT, ANHYDROUS_MOLECULAR_WEIGHT_CONSTANT
 from .params import BASE_EXTINCTION_COEFFICIENTS, NN53_EXTINCTION_COEFFICIENTS
-from .params import CODONS_TO_AMINO_ACIDS
 from .params import NN_PARAMS, DNA_COMPLEMENT_MAP
+from .params import CODONS_TO_AMINO_ACIDS_SHORT, CODONS_TO_AMINO_ACIDS_LONG
 
 def molecular_weight_calc(sequence):
     """
@@ -151,8 +151,7 @@ def e260_ssnn_calc(sequence):
         e260 -= BASE_EXTINCTION_COEFFICIENTS[base]
     return e260
 
-
-def protein_seq_calc(rna_sequence, frame):
+def protein_seq_calc(rna_sequence, frame, multi_letter=False):
     """
     Calculate the sequence of amino acids from a mRNA sequence.
 
@@ -160,12 +159,14 @@ def protein_seq_calc(rna_sequence, frame):
     :type rna_sequence: str
     :param frame: frame of translation (1, 2, or 3)
     :type frame: int
+    :param multi_letter: whether to return amino acids in 1-letter codes (False) or 3-letter codes (True)
+    :type multi_letter: bool
     :return: protein sequence
     """
-    frame = frame - 1
-    protein_seq = ""
-    for i in range(frame, len(rna_sequence), 3):
-        codon = rna_sequence[i:i + 3]
-        if len(codon) == 3:
-            protein_seq += CODONS_TO_AMINO_ACIDS[codon]
-    return protein_seq
+    start = frame - 1
+    protein = []
+    codon_to_amino_acid = CODONS_TO_AMINO_ACIDS_LONG if multi_letter else CODONS_TO_AMINO_ACIDS_SHORT
+    for i in range(start, len(rna_sequence) - 2, 3):
+        codon = rna_sequence[i:i+3]
+        protein.append(codon_to_amino_acid[codon])
+    return '-'.join(protein) if multi_letter else ''.join(protein)