+++ CHANGELOG.md
@@ -5,11 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `to_protein` method
 ### Changed
 - Test system modified
 - `Python 3.6` support dropped
-### Added
-- `to_protein` method
 ## [0.4] - 2025-03-18
 ### Added
 - Salt-adjusted melting temperature calculation