+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Test system modified
 - `Python 3.6` support dropped
+- `README.md` updated
 ## [0.4] - 2025-03-18
 ### Added
 - Salt-adjusted melting temperature calculation