+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Thermodynamic constants (ΔH, ΔS) calculation
 - Nearest neighbor melting temperature calculation
 - `to_protein` method
+- `protein_seq_calc` function
 ### Changed
 - Test system modified
 - `Python 3.6` support dropped